+++ tests/test_avx2_div.c
@@ -11,7 +11,7 @@ main(void)
   int s;
   compute_magic_div(divisor, &m, &s);
 
-  printf("Divisor %u: m=%lu, s=%d\n", divisor, m, s);
+  printf("Divisor %u: m=%llu, s=%d\n", divisor, (unsigned long long)m, s);
 
   // Test with values: 40, 41, 42, 43
   __m256i test_values = _mm256_set_epi64x(43, 42, 41, 40);
@@ -21,14 +21,14 @@ main(void)
   _mm256_storeu_si256((__m256i*)results, result);
 
   printf("Input: [40, 41, 42, 43]\n");
-  printf("Output: [%lu, %lu, %lu, %lu]\n", results[0], results[1], results[2], results[3]);
+  printf("Output: [%llu, %llu, %llu, %llu]\n", (unsigned long long)results[0], (unsigned long long)results[1], (unsigned long long)results[2], (unsigned long long)results[3]);
   printf("Expected: [4, 4, 4, 4]\n");
 
   int errors = 0;
   uint64_t expected[] = {4, 4, 4, 4};
   for (int i = 0; i < 4; ++i) {
     if (results[i] != expected[i]) {
-      printf("ERROR at index %d: expected %lu, got %lu\n", i, expected[i], results[i]);
+      printf("ERROR at index %d: expected %llu, got %llu\n", i, (unsigned long long)expected[i], (unsigned long long)results[i]);
       errors++;
     }
   }
